+++ scripts/sidecar-chrome.mjs
@@ -167,6 +167,7 @@ function startChrome() {
 		"  --no-sandbox \\",
 		"  --password-store=basic \\",
 		"  --hide-crash-restore-bubble \\",
+		"  --js-flags=--max-old-space-size=1536 \\",
 		"  --simulate-outdated-no-au='Tue, 31 Dec 2099 23:59:59 GMT' \\",
 		"  --start-maximized \\",
 		"  --test-type \\",